// src/types.ts
/**
 * Copyright (c) Ladifire, Inc. and its affiliates.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */

export interface AnimationProps {
    frameCount: number;
    framesPerCol: number;
    framesPerRow: number;
    step: number;
}